--- a/test/e2e/correlation_e2e_test.go
+++ b/test/e2e/correlation_e2e_test.go
@@ -2,7 +2,6 @@ package e2e
 
 import (
 	"encoding/json"
-	"fmt"
 	"os"
 	"os/exec"
 	"path/filepath"