… .github/workflows/cache-cleanup.yml
… .github/workflows/ci.yml
… .github/workflows/comprehensive-tests.yml
… .github/workflows/release.yaml
… .github/workflows/selective-tests.yml
… .goreleaser.yml
… CLAUDE.md
… cmd/wgo/commands/diff.go
… cmd/wgo/commands/help.go
… cmd/wgo/commands/root.go
… cmd/wgo/commands/scan.go
… cmd/wgo/commands/status.go
… coverage.html
… coverage.out
… docs/INSTALLATION.md
… docs/ci-configuration.md
… docs/testing-strategy.md
… go.mod
… go.sum
… internal/app/factory.go
… internal/collectors/aws/README.md
… internal/collectors/aws/aws_system_test.go
… internal/collectors/aws/aws_test.go
… internal/collectors/aws/client.go
… internal/collectors/aws/client_test.go
… internal/collectors/aws/collector_test.go
… internal/collectors/aws/ec2_test.go
… internal/collectors/aws/mocks.go
… internal/collectors/aws/normalizer_test.go
… internal/collectors/aws/testdata/aws_mock_response.json